Refactor: Enhance AI Recommendation Settings and Implement OpenAI Provider
This commit refactors the AI settings UI and implements the OpenAI provider for recommendations.

- **AI Settings UI Refactoring:**
    - `kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/recommendations/RecommendationScreen.kt`:
        - Removed the inline `AiSettings`, `GeminiSettings`, and `OpenAiSettings` composables.
        - The main `Scaffold` in `RecommendationScreen` now uses `Modifier.fillMaxSize()`.
        - The `BottomAppBar` now uses `windowInsetsPadding(BottomAppBarDefaults.windowInsets)` instead of `navigationBarsPadding()`.
        - Removed `imePadding()` from the inner `Scaffold`.
    - `kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/recommendations/AiSettings.kt`:
        - Created a new file to house the `AiSettings`, `GeminiSettings`, and `OpenAiSettings` composables.
        - `AiSettings` now uses a `Scaffold` with a `BottomAppBar` containing the "Save" button, which respects window insets.
        - Added `Crossfade` animation when switching between AI service settings (Gemini/OpenAI).
        - API key fields in `GeminiSettings` and `OpenAiSettings` now use `PasswordVisualTransformation` by default and can be toggled to show the key (though the toggle functionality is not explicitly shown in this diff but implied by `showApiKey` variable).

- **OpenAI Provider Implementation:**
    - `kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/recommendations/aiproviders/OpenAiProvider.kt`:
        - Implemented the `OpenAiProvider`.
        - The `init()` function initializes an `AiClient` (from `io.github.bay73:generic-ai`) for OpenAI using the API key and model name from `newSettingsHandling.aiSettings`.
        - The `getResult(prompt: String)` function now uses the initialized `AiClient` to generate text based on the provided prompt and system instructions from settings. It handles potential failures and returns the response or error message.

- **Dependency Updates:**
    - `build.gradle.kts`:
        - Updated `org.jetbrains.compose.hot-reload` plugin version from `1.0.0-alpha10` to `1.0.0-alpha11`.
    - `kmpuiviews/build.gradle.kts`:
        - Added `implementation("io.github.bay73:generic-ai:0.6.5")` to commonMain dependencies for the OpenAI client.
